Refactor project to use a docker container
Add a Dockerfile that builds an up to date image with all of our
dependencies that is based on our build-base image. The contents are
copied here so to make it easier to transition this to an open source
repository that is available to our OSS projects as well.

The logic of the SDCLI commands is translated to bash scripts with a
minimal structure for adding new supported tech stacks and new commands
for go.

The "check" series of commands is not included here because it is not
possible to run the checks inside a container. They must be executed on
a host. This feature should spin off to a new repository that focuses on
helping new devs onboard.