Add initial batch of repo management commands
These commands would replace direct usage of our cookiecutter templates
with calls to sdcli. A command is also added to audit whether or not the
current repo implements the contract. We should likely add the audit
command to our CI/CD pipelines to enforce that projects comply without
the need for a centralized test that checks all the repos.

This patch also makes the suggestion that we should refactor our
templates to be much more granular rather than a single template for a
project. Doing this would enable us to incrementally update smaller
components, such as the linting configuration or pipelines files,
without folks needing to run the full project render.